+++ examples/confirm.rs
@@ -1,7 +1,7 @@
-use dialoguer::Confirm;
+use dialoguer::{theme::ColorfulTheme, Confirm};
 
 fn main() {
-    if Confirm::new()
+    if Confirm::with_theme(&ColorfulTheme::default())
         .with_prompt("Do you want to continue?")
         .interact()
         .unwrap()
@@ -9,27 +9,40 @@ fn main() {
         println!("Looks like you want to continue");
     } else {
         println!("nevermind then :(");
-        return;
     }
 
-    println!("disable default");
-    if Confirm::new()
-        .with_prompt("continue?")
-        .disable_default(true)
+    if Confirm::with_theme(&ColorfulTheme::default())
+        .with_prompt("Do you really want to continue?")
+        .default(true)
         .interact()
         .unwrap()
     {
-        println!("continuing");
+        println!("Looks like you want to continue");
     } else {
-        println!("exiting");
+        println!("nevermind then :(");
     }
 
-    println!();
-
-    println!("enable default");
-    if Confirm::new().with_prompt("continue?").interact().unwrap() {
-        println!("continuing");
+    if Confirm::with_theme(&ColorfulTheme::default())
+        .with_prompt("Do you really really want to continue?")
+        .default(true)
+        .show_default(false)
+        .wait_for_newline(true)
+        .interact()
+        .unwrap()
+    {
+        println!("Looks like you want to continue");
     } else {
-        println!("exiting");
+        println!("nevermind then :(");
+    }
+
+    if Confirm::with_theme(&ColorfulTheme::default())
+        .with_prompt("Do you really really really want to continue?")
+        .wait_for_newline(true)
+        .interact()
+        .unwrap()
+    {
+        println!("Looks like you want to continue");
+    } else {
+        println!("nevermind then :(");
     }
 }