+++ src/prompts/confirm.rs
@@ -21,9 +21,8 @@ use console::Term;
 /// ```
 pub struct Confirm<'a> {
     prompt: String,
-    default: bool,
+    default: Option<bool>,
     show_default: bool,
-    disable_default: bool,
     wait_for_newline: bool,
     theme: &'a dyn Theme,
 }
@@ -59,9 +58,8 @@ impl<'a> Confirm<'a> {
     pub fn with_theme(theme: &'a dyn Theme) -> Confirm<'a> {
         Confirm {
             prompt: "".into(),
-            default: true,
+            default: None,
             show_default: true,
-            disable_default: false,
             wait_for_newline: false,
             theme,
         }
@@ -92,32 +90,24 @@ impl<'a> Confirm<'a> {
         self.wait_for_newline = wait;
         self
     }
-    /// Overrides the default output if user pushes enter key without inputing any character.
-    /// Character corresponding to the default choice (e.g `Y` if default is `true`) will be uppercased in the displayed prompt.
+
+    /// Sets a default.
     ///
-    /// The default output is true.
+    /// Out of the box the prompt does not have a default and will continue
+    /// to display until the user inputs something and hits enter. If a default is set the user
+    /// can instead accept the default with enter.
     pub fn default(&mut self, val: bool) -> &mut Confirm<'a> {
-        self.default = val;
+        self.default = Some(val);
         self
     }
 
-    /// Disables or enables display of options user can choose from.
+    /// Disables or enables the default value display.
     ///
     /// The default is to append `[y/n]` to the prompt to tell the
     /// user which keys to press. This also renders the default choice
     /// in uppercase. The default is selected on enter.
     pub fn show_default(&mut self, val: bool) -> &mut Confirm<'a> {
         self.show_default = val;
-        self
-    }
-
-    /// Select whether the user must explicitly confirm their selection.
-    ///
-    /// When `false` (default), the user can input a newline to select the default.
-    ///
-    /// When `true`, the user must input a letter - a newline will do nothing.
-    pub fn disable_default(&mut self, val: bool) -> &mut Confirm<'a> {
-        self.disable_default = val;
         self
     }
 
@@ -149,60 +139,76 @@ impl<'a> Confirm<'a> {
     pub fn interact_on(&self, term: &Term) -> io::Result<bool> {
         let mut render = TermThemeRenderer::new(term, self.theme);
 
-        let default = if self.show_default {
-            Some(self.default)
+        let default_if_show = if self.show_default {
+            self.default
         } else {
             None
         };
-        render.confirm_prompt(&self.prompt, default)?;
+
+        render.confirm_prompt(&self.prompt, default_if_show)?;
 
         term.hide_cursor()?;
         term.flush()?;
 
+        let rv;
+
         if self.wait_for_newline {
             // Waits for user input and for the user to hit the Enter key
             // before validation.
-            let mut input_buf = String::new();
-            loop {
-                io::stdin().read_line(&mut input_buf)?;
-                let rv = match &*input_buf.trim_end().to_lowercase() {
-                    "y" | "yes" => true,
-                    "n" | "no" => false,
-                    "" if !self.disable_default => self.default,
-                    _ => {
-                        // On invalid input re-render the user prompt.
-                        render.confirm_prompt(&self.prompt, default)?;
-                        input_buf.clear();
-                        continue;
-                    }
-                };
+            let mut value = default_if_show;
 
-                term.show_cursor()?;
-                term.flush()?;
-
-                return Ok(rv);
-            }
-        } else {
-            // Default behavior: matches continuously on every keystroke,
-            // and does not wait for user to hit the Enter key.
             loop {
                 let input = term.read_char()?;
-                let rv = match input {
-                    'y' | 'Y' => true,
-                    'n' | 'N' => false,
-                    '\n' | '\r' if !self.disable_default => self.default,
+
+                match input {
+                    'y' | 'Y' => {
+                        value = Some(true);
+                    }
+                    'n' | 'N' => {
+                        value = Some(false);
+                    }
+                    '\n' | '\r' => {
+                        value = value.or(self.default);
+
+                        if let Some(val) = value {
+                            rv = val;
+                            break;
+                        } else {
+                            continue;
+                        }
+                    }
                     _ => {
                         continue;
                     }
                 };
 
                 term.clear_line()?;
-                render.confirm_prompt_selection(&self.prompt, rv)?;
-                term.show_cursor()?;
-                term.flush()?;
+                render.confirm_prompt(&self.prompt, value)?;
+            }
+        } else {
+            // Default behavior: matches continuously on every keystroke,
+            // and does not wait for user to hit the Enter key.
+            loop {
+                let input = term.read_char()?;
+                let value = match input {
+                    'y' | 'Y' => true,
+                    'n' | 'N' => false,
+                    '\n' | '\r' if self.default.is_some() => self.default.unwrap(),
+                    _ => {
+                        continue;
+                    }
+                };
 
-                return Ok(rv);
+                rv = value;
+                break;
             }
         }
+
+        term.clear_line()?;
+        render.confirm_prompt_selection(&self.prompt, rv)?;
+        term.show_cursor()?;
+        term.flush()?;
+
+        return Ok(rv);
     }
 }