+++ CHANGELOG.md
@@ -5,7 +5,6 @@
 ### Enhancements
 
 * Added `wait_for_newline` to `Confirm`
-* Added `disable_default` to `Confirm`
 * More secure password prompt
 * More documentation
 * Added `interact_text` method for input prompt
@@ -15,6 +14,7 @@
 
 * Removed `theme::CustomPromptCharacterTheme`
 * `Input` validators now take the input type `T` as arg
+* `Confirm` has no `default` value by default now
 
 ## 0.6.2